+++ src/core/execution.unit.handlers.ts
@@ -1,4 +1,4 @@
-import { rest } from 'msw'
+import { HttpResponse, http } from 'msw'
 import { buildStepObject } from '../../tests/fixtures.js'
 import { config } from '../config.js'
 import {
@@ -10,29 +10,21 @@ import {
 const _config = config.get()
 
 export const lifiHandlers = [
-  rest.post(
-    `${_config.apiUrl}/advanced/stepTransaction`,
-    async (_, response, context) =>
-      response(
-        context.status(200),
-        context.json(
-          mockStepTransactionWithTxRequest(
-            buildStepObject({
-              includingExecution: true,
-            })
-          )
-        )
+  http.post(`${_config.apiUrl}/advanced/stepTransaction`, async () =>
+    HttpResponse.json(
+      mockStepTransactionWithTxRequest(
+        buildStepObject({
+          includingExecution: true,
+        })
       )
-  ),
-  rest.get(`${_config.apiUrl}/chains`, async (_, response, context) =>
-    response(
-      context.status(200),
-      context.json({
-        chains: mockChainsResponse,
-      })
     )
   ),
-  rest.get(`${_config.apiUrl}/status`, async (_, response, context) =>
-    response(context.status(200), context.json(mockStatus))
+  http.get(`${_config.apiUrl}/chains`, async () =>
+    HttpResponse.json({
+      chains: mockChainsResponse,
+    })
+  ),
+  http.get(`${_config.apiUrl}/status`, async () =>
+    HttpResponse.json(mockStatus)
   ),
 ]