+++ vitest.config.ts
@@ -1,10 +1,9 @@
-import tsconfigPaths from 'vite-tsconfig-paths'
+// import tsconfigPaths from 'vite-tsconfig-paths'
 import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
-  plugins: [tsconfigPaths() as any],
+  // plugins: [tsconfigPaths() as any],
   test: {
-    setupFiles: ['./tests/setupEnv.ts'],
     dangerouslyIgnoreUnhandledErrors: true,
   },
 })