+++ vitest.e2e.config.ts
@@ -2,7 +2,6 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
-    setupFiles: ['./tests/setupEnv.ts'],
     dangerouslyIgnoreUnhandledErrors: true,
     testTimeout: 999_999_999,
     hookTimeout: 999_999_999,