+++ CHANGELOG.md
@@ -2,6 +2,40 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.5.0](https://github.com/lifinance/sdk/compare/v2.4.3...v2.5.0) (2023-11-01)
+
+
+### Features
+
+* upgrade to msw v2 ([f1cde15](https://github.com/lifinance/sdk/commit/f1cde15326cf10e9ae4c39b08603fb181fdbacd2))
+
+
+### Bug Fixes
+
+* improve error handling during LiFi initialization ([#171](https://github.com/lifinance/sdk/issues/171)) ([0756631](https://github.com/lifinance/sdk/commit/075663180b50383a8672bb987dd841fd4ac3c43b))
+
+### [2.4.3](https://github.com/lifinance/sdk/compare/v2.4.1...v2.4.3) (2023-10-03)
+
+
+### Bug Fixes
+
+* wallet analytics API request changes ([#167](https://github.com/lifinance/sdk/issues/167)) ([e4a145d](https://github.com/lifinance/sdk/commit/e4a145de96ef986b8ea52134cfb095d4a5fc5f44))
+
+### [2.4.2](https://github.com/lifinance/sdk/compare/v2.4.1...v2.4.2) (2023-10-03)
+
+
+### Bug Fixes
+
+* wallet analytics API request changes ([#167](https://github.com/lifinance/sdk/issues/167)) ([e4a145d](https://github.com/lifinance/sdk/commit/e4a145de96ef986b8ea52134cfb095d4a5fc5f44))
+
+### [2.4.1](https://github.com/lifinance/sdk/compare/v2.4.0...v2.4.1) (2023-09-25)
+
+
+### Bug Fixes
+
+* publish workflow ([c005f0a](https://github.com/lifinance/sdk/commit/c005f0a56a4df56ceceef90b4a220964094b1056))
+* type fix for transaction history response ([#166](https://github.com/lifinance/sdk/issues/166)) ([ee954b8](https://github.com/lifinance/sdk/commit/ee954b875f8a865b9504dd714f5dceb36b862d73))
+
 ## [2.4.0](https://github.com/lifinance/sdk/compare/v2.3.1...v2.4.0) (2023-09-20)